+++ Nuspec4NuspecModule.ps1
@@ -1,5 +1,7 @@
 ﻿
-nuspec 'PSNuspec' '0.2.0' {
+$ModuleVersion=(Import-ManifestData "$NuspecVcs\PSNuspec.psd1").ModuleVersion
+
+nuspec 'PSNuspec' $ModuleVersion {
    properties @{
         Authors='Dardenne Laurent'
         Owners='Dardenne Laurent'
@@ -11,13 +13,7 @@ nuspec 'PSNuspec' '0.2.0' {
         licenseUrl='https://creativecommons.org/licenses/by-nc-sa/4.0/'
         projectUrl='https://github.com/LaurentDardenne/Nuspec'
         iconUrl='https://github.com/LaurentDardenne/Nuspec/blob/master/icon/Nuspec.png'
-        releaseNotes=@'
-2016-09-13  Version 0.2.0
-Use the latest XSD file (Nuget.exe v3.5 rc1)
-
-2016-09-11  Version 0.1.0.0
-Original version
-'@
+        releaseNotes="$(Get-Content "$NuspecVcs\CHANGELOG.md" -raw)"
         tags='PSModule PSIncludes_Function PSFunction_nuspec PSFunction_properties PSFunction_Dependencies PSFunction_dependency PSFunction_Files PSFunction_file PSFunction_Save-Nuspec PSCommand_nuspec PSCommand_properties PSCommand_Dependencies PSCommand_dependency PSCommand_Files PSCommand_file PSCommand_Save-Nuspec'
         #tags='nuspec XSD Powershell DSL'
    }
@@ -28,7 +24,7 @@ Original version
    
    files {
         file -src "$NuspecVcs\lib\net40\NugetSchema.dll" -target "lib\net40\NugetSchema.dll"
-        file -src "$NuspecVcs\nuspec.2011.8.xsd"
+        file -src "$NuspecVcs\nuspec.2013.05.xsd"
         file -src "$NuspecVcs\PSNuspec.psd1"
         file -src "$NuspecVcs\PSNuspec.psm1"
         file -src "$NuspecVcs\README.md"