+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-﻿2016-09-13  Version 0.2.0.0
+﻿2016-09-14  Version 0.3.0
+Fix the $Stack scope
+Fix the namespace name
+Add functions Import-ManifestData, Get-ScriptVersion 
+
+
+2016-09-13  Version 0.2.0
 Use the latest XSD file (Nuget.exe v3.5 rc1)
 
 2016-09-11  Version 0.1.0.0